+++ SRC/NanoSpec/RecipeDistanceStageProgramView.cpp
@@ -1292,7 +1292,7 @@ BOOL CRecipeDistanceStageProgramView::SaveRecipeData()
 			BOOL bCreateDir = TRUE;
 			_fullpath(fullPath,l_UserData.BackupData.BackupPath,sizeof(fullPath));
 			_tcscat(fullPath,"\\" DB_DIR_NAME "\\" DB_DESKEW_IMG_DIR_NAME "\\" );
-			for(int n = 0;n < strlen(fullPath);n++){
+			for(size_t n = 0;n < strlen(fullPath);n++){
 				if(n > 0 && fullPath[n] == '\\' && fullPath[n-1] != ':'){
 					fullPath[n] = '\0';
 					bCreateDir = CreateDirectory(fullPath,NULL);
